Updated binary protocol version to 3.2.71. (#111)
* Updated binary protocol version to 3.2.71.
Added VANILLA_KEYWORD origin.
Don't insert selections if limit exceeded.
Added lock symobol to completions popup and status bar icon.

* Fixed icon in 'limit exceeded' popup.

* Changed 'State' field type in HoverActionRequest (and response) to be of tyoe Object to handle any value.
Added a 'limited changed' topic instead of using the BinaryStateService (Using the service caused an issue with tests).
Changed TestData queries to use the current protocol version instead of a hard coded one.

* Remove the locked icon from the status bar if service level changes to pro.

* Removed unused variable.

* Refactored completions classes for better readability.

* Changed binary state polling frequency.

* Bug fix: don't attempt to shutdown app thread pool.

* Hover response can be null.

* Addeed some comments and renamed methods for better readability.

+++ src/main/java/com/tabnine/intellij/completions/TabNineCompletionContributor.java
@@ -1,10 +1,12 @@
-package com.tabnine;
+package com.tabnine.intellij.completions;
 
 import com.intellij.codeInsight.completion.*;
 import com.intellij.codeInsight.lookup.*;
+import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.editor.Document;
 import com.intellij.openapi.util.TextRange;
+import com.intellij.util.messages.MessageBus;
 import com.tabnine.binary.requests.autocomplete.AutocompleteResponse;
 import com.tabnine.binary.requests.autocomplete.ResultEntry;
 import com.tabnine.config.Config;
@@ -16,6 +18,7 @@ import com.tabnine.prediction.TabNinePrefixMatcher;
 import com.tabnine.prediction.TabNineWeigher;
 import com.tabnine.selections.TabNineLookupListener;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.ArrayList;
 import java.util.Objects;
@@ -26,6 +29,8 @@ import static com.tabnine.general.Utils.endsWithADot;
 public class TabNineCompletionContributor extends CompletionContributor {
     private final CompletionFacade completionFacade = DependencyContainer.instanceOfCompletionFacade();
     private final TabNineLookupListener tabNineLookupListener = DependencyContainer.instanceOfTabNineLookupListener();
+    private final MessageBus messageBus = ApplicationManager.getApplication().getMessageBus();
+    private boolean isLocked;
 
     @Override
     public void fillCompletionVariants(@NotNull CompletionParameters parameters, @NotNull CompletionResultSet resultSet) {
@@ -41,6 +46,12 @@ public class TabNineCompletionContributor extends CompletionContributor {
         if (originalMatcher.getPrefix().length() == 0 && completions.results.length == 0) {
             return;
         }
+        if (this.isLocked != completions.is_locked) {
+            this.isLocked = completions.is_locked;
+            this.messageBus
+              .syncPublisher(LimitedSecletionsChangedNotifier.LIMITED_SELECTIONS_CHANGED_TOPIC)
+              .limitedChanged(completions.is_locked);
+        }
 
         resultSet = resultSet.withPrefixMatcher(new TabNinePrefixMatcher(originalMatcher.cloneWithPrefix(completions.old_prefix)))
                 .withRelevanceSorter(CompletionSorter.defaultSorter(parameters, originalMatcher).weigh(new TabNineWeigher()));
@@ -53,19 +64,24 @@ public class TabNineCompletionContributor extends CompletionContributor {
 
     private ArrayList<LookupElement> createCompletions(AutocompleteResponse completions, @NotNull CompletionParameters parameters, @NotNull CompletionResultSet resultSet) {
         ArrayList<LookupElement> elements = new ArrayList<>();
+        final Lookup activeLookup = LookupManager.getActiveLookup(parameters.getEditor());
+        for (int index = 0; index < completions.results.length && index < completionLimit(parameters, resultSet, completions.is_locked); index++) {
+            LookupElement lookupElement = createCompletion(
+                    parameters, resultSet, completions.old_prefix,
+                    completions.results[index], index, completions.is_locked, activeLookup);
 
-        for (int index = 0; index < completions.results.length && index < completionLimit(parameters, resultSet); index++) {
-            LookupElementBuilder lookupElementBuilder = createCompletion(parameters, resultSet, completions.old_prefix, completions.results[index], index);
-
-            if (resultSet.getPrefixMatcher().prefixMatches(lookupElementBuilder)) {
-                elements.add(lookupElementBuilder);
+            if (resultSet.getPrefixMatcher().prefixMatches(lookupElement)) {
+                elements.add(lookupElement);
             }
         }
 
         return elements;
     }
 
-    private int completionLimit(CompletionParameters parameters, CompletionResultSet result) {
+    private int completionLimit(CompletionParameters parameters, CompletionResultSet result, boolean isLocked) {
+        if (isLocked) {
+            return 1;
+        }
         boolean preferTabNine = !endsWithADot(
                 parameters.getEditor().getDocument(),
                 parameters.getOffset() - result.getPrefixMatcher().getPrefix().length()
@@ -75,8 +91,9 @@ public class TabNineCompletionContributor extends CompletionContributor {
     }
 
     @NotNull
-    private LookupElementBuilder createCompletion(CompletionParameters parameters, CompletionResultSet resultSet,
-                                                  String oldPrefix, ResultEntry result, int index) {
+    private LookupElement createCompletion(CompletionParameters parameters, CompletionResultSet resultSet,
+                                                  String oldPrefix, ResultEntry result, int index,
+                                           boolean locked, @Nullable Lookup activeLookup) {
         TabNineCompletion completion = new TabNineCompletion(
                 oldPrefix,
                 result.new_prefix,
@@ -95,35 +112,49 @@ public class TabNineCompletionContributor extends CompletionContributor {
             completion.deprecated = result.deprecated;
         }
 
-        return LookupElementBuilder.create(completion, result.new_prefix)
-                .withInsertHandler((context, item) -> {
-                    int end = context.getTailOffset();
-                    TabNineCompletion lookupElement = (TabNineCompletion) item.getObject();
-                    try {
-                        context.getDocument().insertString(end + lookupElement.oldSuffix.length(), lookupElement.newSuffix);
-                        context.getDocument().deleteString(end, end + lookupElement.oldSuffix.length());
-                    } catch(RuntimeException re) {
-                        Logger.getInstance(getClass()).warn("Error inserting new suffix. End = " + end +
-                                ", old suffix length = " + lookupElement.oldSuffix.length() + ", new suffix length = "
-                                + lookupElement.newSuffix.length(), re);
+        LookupElementBuilder lookupElementBuilder =
+            LookupElementBuilder.create(completion, result.new_prefix)
+            .withRenderer(
+                new LookupElementRenderer<LookupElement>() {
+                  @Override
+                  public void renderElement(
+                      LookupElement element, LookupElementPresentation presentation) {
+                    TabNineCompletion lookupElement = (TabNineCompletion) element.getObject();
+                    String typeText = (locked ? LIMITATION_SYMBOL : "");
+                    if (Config.DISPLAY_ORIGIN) {
+                      typeText += lookupElement.origin.toString();
+                    } else {
+                      typeText += StaticConfig.BRAND_NAME;
                     }
-                }).withRenderer(new LookupElementRenderer<LookupElement>() {
-                    @Override
-                    public void renderElement(LookupElement element, LookupElementPresentation presentation) {
-                        TabNineCompletion lookupElement = (TabNineCompletion) element.getObject();
-
-                        if (Config.DISPLAY_ORIGIN) {
-                            presentation.setTypeText(lookupElement.origin.toString());
-                        } else {
-                            presentation.setTypeText(StaticConfig.BRAND_NAME);
-                        }
-
-                        presentation.setItemTextBold(false);
-                        presentation.setStrikeout(lookupElement.deprecated);
-                        presentation.setItemText(lookupElement.newPrefix);
-                        presentation.setIcon(ICON);
-                    }
+                    presentation.setTypeText(typeText);
+                    presentation.setItemTextBold(false);
+                    presentation.setStrikeout(lookupElement.deprecated);
+                    presentation.setItemText(lookupElement.newPrefix);
+                    presentation.setIcon(ICON);
+                  }
                 });
+        if (locked) {
+            final InsertNothingLookupElement lookupElement = new LimitExceededLookupElement(
+                    lookupElementBuilder, oldPrefix);
+            if (activeLookup != null) {
+                activeLookup.addLookupListener(lookupElement);
+            }
+            return lookupElement;
+        } else {
+            lookupElementBuilder = lookupElementBuilder.withInsertHandler((context, item) -> {
+                int end = context.getTailOffset();
+                TabNineCompletion lookupElement = (TabNineCompletion) item.getObject();
+                try {
+                    context.getDocument().insertString(end + lookupElement.oldSuffix.length(), lookupElement.newSuffix);
+                    context.getDocument().deleteString(end, end + lookupElement.oldSuffix.length());
+                } catch(RuntimeException re) {
+                    Logger.getInstance(getClass()).warn("Error inserting new suffix. End = " + end +
+                            ", old suffix length = " + lookupElement.oldSuffix.length() + ", new suffix length = "
+                            + lookupElement.newSuffix.length(), re);
+                }
+            });
+        }
+        return lookupElementBuilder;
     }
 
     private void addAdvertisement(@NotNull CompletionResultSet result, AutocompleteResponse completions) {
